Fixed LastThrouhgput To Fix NoOp in MultiRate Situations
Need to track the throughput of the last node to know how parallel
a noop should be. It's parallel is not dependent on the current
throuhgput as it isn't a real node, it's just a pass through
of a prior node. This ensures the tracking occurs in the right
place, in the appendToCompilationData function rather than in
>>>. Also, it fixes the sseq to vec and tuple converters to not
modiy last throughput as they are not changing throuhgput,
their making atoms.